refactor(sdk): extract shared error proxy fixture to conftest.py
